+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # developing
+- [add non-block write log in asynchronous mode](https://github.com/beego/beego/pull/5150)
 - [Fix 5126: support bloom filter cache](https://github.com/beego/beego/pull/5126)
 - [Fix 5117: support write though cache](https://github.com/beego/beego/pull/5117)
 - [add read through for cache module](https://github.com/beego/beego/pull/5116)